hit a wall trying to extract bibliography section
Extracting the bibliography section is easy enough if there is a
heading like "References". If not, my heuristics (does the line
contain a four-digit number? does it contain lots of upper-case
letters? etc.) didn't manage to reliably retrieve the section.

This is a problem I already encountered with author_score: the best
way to assign a score is to try extracting the relevant information.

I will try to rewrite the Extractor so that block separation goes hand
in hand with block tagging and metadata extraction: a block that that
looks like a citation often continues over indented lines, while
paragraph blocks don't.